+++ stories/components/dropdown-menu/docs/Dropdown-menu.stories.tsx
@@ -9,6 +9,7 @@ import { MultiselectItemsTest } from './__tests__/multiselect_items.test';
 import { NestedTest } from './__tests__/nested.test';
 import { NestedWithFocusableTest } from './__tests__/nested_with_focusable.test';
 import { SelectableRadioTest } from './__tests__/selectable_radio_items.test';
+import { StickyGroupTest } from './__tests__/sticky_group.test';
 import { TheSecondMethodTest } from './__tests__/the_second_method.test';
 import BasicExample from './examples/basic';
 import DropdownMenuExample from './examples/dropdown-menu';
@@ -18,6 +19,7 @@ import MultiselectItemsExample from './examples/multiselect_items';
 import NestedExample from './examples/nested';
 import NestedWithFocusableExample from './examples/nested_with_focusable';
 import SelectableRadioItemsExample from './examples/selectable_radio_items';
+import StickyGroupsExample from './examples/sticky_groups';
 import TheSecondMethodExample from './examples/the_second_method';
 import { playWrapper } from '../../../utils/playWrapper';
 
@@ -72,4 +74,9 @@ export const SelectableRadioItems: Story = {
 export const MultiselectItems: Story = {
   render: MultiselectItemsExample,
   play: playWrapper(MultiselectItemsTest),
+};
+
+export const StickyGroups: Story = {
+  render: StickyGroupsExample,
+  play: playWrapper(StickyGroupTest),
 };