+++ semcore/dropdown/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.3] - 2025-06-20
+
+### Added
+
+- `sticky` property for Dropdown.Group.
+
+### Change
+
+- Clear highlightedIndex after closing popover to correct default value depend on type - `0` for DropdownMenu and `null` for Select.
+
 ## [16.0.2] - 2025-06-12
 
 ### Fixed