+++ semcore/tooltip/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.0.3] - 2025-06-16
+## [16.0.3] - 2025-06-23
 
-### Fixed
+### Changed
 
-- Page layout appeared shifted due to width/height changes.
+- Version patch update due to children dependencies update (`@semcore/flex-box` [16.0.2 ~> 16.0.3], `@semcore/popper` [16.0.2 ~> 16.0.3], `@semcore/portal` [16.0.2 ~> 16.0.3], `@semcore/base-components` [16.0.2 ~> 16.1.0]).
 
 ## [16.0.2] - 2025-06-16